+++ rfrx/protronik.py
@@ -7,25 +7,25 @@ from .sbus import SbusDecoder, SbusReader
 
 LOGGER = getLogger("rfrx.protronik")
 
-RH_MIN, RH_MID, RH_MAX = (
-    int(environ.get("RFRX_RH_MIN", 192)),
-    int(environ.get("RFRX_RH_MID", 992)),
-    int(environ.get("RFRX_RH_MAX", 1796)),
+RX_MIN, RX_MID, RX_MAX = (
+    int(environ.get("RFRX_RX_MIN", 192)),
+    int(environ.get("RFRX_RX_MID", 992)),
+    int(environ.get("RFRX_RX_MAX", 1796)),
 )
-RV_MIN, RV_MID, RV_MAX = (
-    int(environ.get("RFRX_RV_MIN", 302)),
-    int(environ.get("RFRX_RV_MID", 1100)),
-    int(environ.get("RFRX_RV_MAX", 1900)),
+RY_MIN, RY_MID, RY_MAX = (
+    int(environ.get("RFRX_RY_MIN", 302)),
+    int(environ.get("RFRX_RY_MID", 1100)),
+    int(environ.get("RFRX_RY_MAX", 1900)),
 )
-LV_MIN, LV_MID, LV_MAX = (
-    int(environ.get("RFRX_LV_MIN", 180)),
-    int(environ.get("RFRX_LV_MID", 980)),
-    int(environ.get("RFRX_LV_MAX", 1779)),
+LY_MIN, LY_MID, LY_MAX = (
+    int(environ.get("RFRX_LY_MIN", 180)),
+    int(environ.get("RFRX_LY_MID", 980)),
+    int(environ.get("RFRX_LY_MAX", 1779)),
 )
-LH_MIN, LH_MID, LH_MAX = (
-    int(environ.get("RFRX_LH_MIN", 192)),
-    int(environ.get("RFRX_LH_MID", 992)),
-    int(environ.get("RFRX_LH_MAX", 1790)),
+LX_MIN, LX_MID, LX_MAX = (
+    int(environ.get("RFRX_LX_MIN", 192)),
+    int(environ.get("RFRX_LX_MID", 992)),
+    int(environ.get("RFRX_LX_MAX", 1790)),
 )
 
 
@@ -37,10 +37,10 @@ class ProTronikDecoder(SbusDecoder):
         super().__init__(frame, n_chans=6)
 
         # calibration: minimum, middle, maximum, name
-        rh = RH_MIN, RH_MID, RH_MAX, "right horizontal"
-        rv = RV_MIN, RV_MID, RV_MAX, "right vertical"
-        lv = LV_MIN, LV_MID, LV_MAX, "left vertical"
-        lh = LH_MIN, LH_MID, LH_MAX, "left horizontal"
+        rx = RX_MIN, RX_MID, RX_MAX, "right horizontal"
+        ry = RY_MIN, RY_MID, RY_MAX, "right vertical"
+        ly = LY_MIN, LY_MID, LY_MAX, "left vertical"
+        lx = LX_MIN, LX_MID, LX_MAX, "left horizontal"
 
         def scale(val, mini, middle, maxi, name):
             """Convert read value to -1.0 -> 1.0."""
@@ -54,10 +54,10 @@ class ProTronikDecoder(SbusDecoder):
                 return (val - mini) / (middle - mini) - 1.0
             return (val - middle) / (maxi - middle)
 
-        self.rh = +scale(self.chans[0], *rh)
-        self.rv = -scale(self.chans[1], *rv)
-        self.lv = -scale(self.chans[2], *lv)
-        self.lh = +scale(self.chans[3], *lh)
+        self.rx = +scale(self.chans[0], *rx)
+        self.ry = -scale(self.chans[1], *ry)
+        self.ly = -scale(self.chans[2], *ly)
+        self.lx = +scale(self.chans[3], *lx)
 
         self.ch5 = 2 if self.chans[4] < 500 else 1 if self.chans[4] < 1500 else 0
         self.ch6 = 0 if self.chans[5] < 1000 else 1
@@ -65,7 +65,7 @@ class ProTronikDecoder(SbusDecoder):
     def __str__(self):
         """Show decoded data."""
         return (
-            f"{self.rh:+.3f} {self.rv:+.3f} {self.lv:+.3f} {self.lh:+.3f} "
+            f"{self.rx:+.3f} {self.ry:+.3f} {self.ly:+.3f} {self.lx:+.3f} "
             f"{self.ch5} {self.ch6}"
         )